refactor(wren): remove DataSource.get_connection — connectors build their own
Every native connector under wren.connector.* now constructs its own
driver connection directly in __init__. The DataSource dispatcher and the
get_<name>_connection helpers it routed to had become a thin layer that
just unpacked typed ConnectionInfo into the driver's connect() call — for
data sources whose connectors already build connections natively (oracle,
bigquery, databricks, trino, clickhouse, snowflake, athena, redshift,
spark, duckdb, datafusion), the dispatcher was never reached.

Inline the remaining four:

- ``PostgresConnector.__init__`` calls ``psycopg.connect(host=..., port=...,
  dbname=..., user=..., password=..., **kwargs)`` directly. Passing a
  ``ConnectionUrl`` form raises ``WrenError(INVALID_CONNECTION_INFO)`` with
  guidance to use ``PostgresConnectionInfo`` instead.
- ``CannerConnector.__init__`` mirrors the same pattern with
  ``dbname=info.workspace``.
- ``MySqlConnector`` / ``DorisConnector`` call ``MySQLdb.connect(
  **_build_mysql_connect_kwargs(info))`` (already supported both typed and
  URL forms via the local helper).
- ``MSSqlConnector.__init__`` calls ``_build_mssql_connection(info)``, a
  new module-local helper that owns the URL-vs-typed branching plus the
  pyodbc connection-string assembly previously living on
  ``DataSourceExtension``.

Drop from ``wren/model/data_source.py``:

- ``DataSource.get_connection`` method.
- The whole ``DataSourceExtension`` Enum, including
  ``_connect_mssql_pyodbc``, ``_escape_odbc_value``,
  ``_register_mssql_output_converters``, ``_decode_mssql_datetimeoffset``,
  ``MSSQL_DATETIMEOFFSET_TYPE_CODE`` (moved to ``wren/connector/mssql.py``)
  and every ``get_<name>_connection`` static method (athena, canner,
  clickhouse, mssql, mysql, doris, postgres, snowflake, trino).
- ``BackendOrConnection`` Union alias and the TYPE_CHECKING imports it
  carried (``MySQLdb``, ``psycopg``, ``PyAthenaConnection``,
  ``TrinoDbApiConnection``).
- Top-level ``datetime`` / ``pyodbc`` / ``urlparse`` / ``unquote_plus`` /
  ``Enum`` imports that became unused once the class was gone.

After this change, ``DataSource`` only carries ``get_connection_info``
(which builds typed ``ConnectionInfo`` objects from request dicts +
applies header-driven timeouts) and ``_build_connection_info`` /
``_handle_clickhouse_url`` — pure data-shape helpers, no driver
involvement.

Tests are updated to import the relocated helpers from
``wren.connector.mssql`` and to use ``AthenaConnector(info)`` directly
where they previously asserted parity between the dispatcher and the
native connector path (that parity is now structurally enforced).

BREAKING CHANGE: ``DataSource.<name>.get_connection(info)`` no longer
exists. Callers should instantiate the connector directly:
``wren.connector.factory.get_connector(data_source, info)`` returns a
``ConnectorABC`` whose ``.connection`` attribute exposes the native
driver handle. ``DataSourceExtension`` is removed from public re-exports.
