+++ core/wren/src/wren/model/data_source.py
@@ -1,15 +1,9 @@
 from __future__ import annotations
 
-import datetime as dtlib
 import urllib
-from enum import Enum, StrEnum, auto
-from typing import TYPE_CHECKING, Any, Union
-from urllib.parse import unquote_plus, urlparse
-
-try:
-    import pyodbc
-except ImportError:  # pragma: no cover
-    pyodbc = None
+from enum import StrEnum, auto
+from typing import Any
+from urllib.parse import unquote_plus
 
 from wren.model import (
     AthenaConnectionInfo,
@@ -40,27 +34,7 @@ from wren.model import (
 )
 from wren.model.error import ErrorCode, WrenError
 
-if TYPE_CHECKING:
-    import MySQLdb
-    import psycopg
-    from pyathena.connection import Connection as PyAthenaConnection
-    from trino.dbapi import Connection as TrinoDbApiConnection
-
-# ``get_connection()`` returns a native driver connection. Each entry below
-# corresponds to one of the ``get_<name>_connection`` static methods on
-# :class:`DataSourceExtension`. Data sources whose native connectors build
-# their connections directly (BigQuery, Oracle, Databricks, Snowflake,
-# ClickHouse, Athena's spark variant, etc.) are not returned through
-# ``get_connection`` — those route through ``wren.connector.factory``.
-BackendOrConnection = Union[
-    "PyAthenaConnection",
-    "MySQLdb.Connection",
-    "psycopg.Connection",
-    "TrinoDbApiConnection",
-]
-
 X_WREN_DB_STATEMENT_TIMEOUT = "x-wren-db-statement_timeout"
-MSSQL_DATETIMEOFFSET_TYPE_CODE = -155
 
 
 class DataSource(StrEnum):
@@ -84,12 +58,6 @@ class DataSource(StrEnum):
     duckdb = auto()
     spark = auto()
     databricks = auto()
-
-    def get_connection(self, info: ConnectionInfo) -> BackendOrConnection:
-        try:
-            return DataSourceExtension[self].get_connection(info)
-        except KeyError:
-            raise NotImplementedError(f"Unsupported data source: {self}")
 
     def get_connection_info(
         self,
@@ -227,372 +195,3 @@ class DataSource(StrEnum):
 
     def _safe_strtobool(self, val: str) -> bool:
         return val.lower() in {"1", "true", "yes", "y"}
-
-
-class DataSourceExtension(Enum):
-    athena = "athena"
-    bigquery = "bigquery"
-    canner = "canner"
-    clickhouse = "clickhouse"
-    datafusion = "datafusion"
-    mssql = "mssql"
-    mysql = "mysql"
-    doris = "doris"
-    oracle = "oracle"
-    postgres = "postgres"
-    redshift = "redshift"
-    snowflake = "snowflake"
-    trino = "trino"
-    local_file = "local_file"
-    duckdb = "duckdb"
-    s3_file = "s3_file"
-    minio_file = "minio_file"
-    gcs_file = "gcs_file"
-    databricks = "databricks"
-    spark = "spark"
-
-    def get_connection(self, info: ConnectionInfo) -> BackendOrConnection:
-        try:
-            if hasattr(info, "connection_url"):
-                # MySQL / Doris use the native MySQLdb driver, not ibis.
-                if self.name in {"mysql", "doris"}:
-                    return getattr(self, f"get_{self.name}_connection")(info)
-                if self.name == "trino":
-                    # Trino uses the native DB-API client; the generic
-                    # ``ibis.connect()`` path was removed when the native
-                    # connector landed. Route the URL through the dedicated
-                    # parser so callers still get a working connection.
-                    from wren.connector.trino import (  # noqa: PLC0415
-                        _build_trino_connect_kwargs,
-                    )
-
-                    trino_kwargs = _build_trino_connect_kwargs(info)
-                    trino_kwargs.pop("_password", None)
-                    trino_kwargs.pop("access_token", None)
-                    from trino.dbapi import (  # noqa: PLC0415
-                        connect as trino_connect,
-                    )
-
-                    return trino_connect(**trino_kwargs)
-                kwargs = info.kwargs if info.kwargs else {}
-                if self.name == "mssql":
-                    return self.get_mssql_connection_from_url(
-                        info.connection_url.get_secret_value(), kwargs
-                    )
-                raise NotImplementedError(
-                    f"connection_url is not supported for {self.name}; "
-                    "use a typed ConnectionInfo instead"
-                )
-            # Data sources whose native connectors construct connections
-            # directly (BigQuery via google-cloud-bigquery, Oracle via oracledb,
-            # Databricks via databricks-sql) and the ones that have no shared
-            # ``get_<name>_connection`` shortcut. Calling
-            # ``DataSource.<name>.get_connection`` for these is unsupported —
-            # construct the typed connector via ``get_connector()`` instead.
-            if self.name in {
-                "bigquery",
-                "databricks",
-                "datafusion",
-                "duckdb",
-                "local_file",
-                "oracle",
-                "redshift",
-                "spark",
-            }:
-                raise NotImplementedError(
-                    f"{self.name}.get_connection() is not supported; "
-                    "use wren.connector.factory.get_connector() instead"
-                )
-            return getattr(self, f"get_{self.name}_connection")(info)
-        except KeyError:
-            raise NotImplementedError(f"Unsupported data source: {self}")
-        except WrenError:
-            raise
-        except Exception as e:
-            raise WrenError(ErrorCode.GET_CONNECTION_ERROR, f"{e!s}") from e
-
-    @staticmethod
-    def get_athena_connection(info: AthenaConnectionInfo) -> PyAthenaConnection:
-        """Open a pyathena DB-API connection.
-
-        Delegates connection-kwargs construction to
-        :func:`wren.connector.athena._build_connect_kwargs` so the legacy
-        ``data_source`` path and the native :class:`AthenaConnector` stay in
-        lockstep on credential resolution, ``schema_name`` propagation,
-        ``kill_on_interrupt``, and user ``kwargs`` merge semantics.
-        """
-        from pyathena import connect  # noqa: PLC0415
-
-        from wren.connector.athena import _build_connect_kwargs  # noqa: PLC0415
-
-        return connect(**_build_connect_kwargs(info))
-
-    @staticmethod
-    def get_canner_connection(info: CannerConnectionInfo):
-        import psycopg  # noqa: PLC0415
-
-        return psycopg.connect(
-            host=info.host,
-            port=int(info.port),
-            dbname=info.workspace,
-            user=info.user,
-            password=info.pat.get_secret_value(),
-            autocommit=True,
-        )
-
-    @staticmethod
-    def get_clickhouse_connection(info: ClickHouseConnectionInfo):
-        import clickhouse_connect  # noqa: PLC0415
-
-        settings = dict(info.settings) if info.settings else {}
-        kwargs = dict(info.kwargs) if info.kwargs else {}
-        statement_timeout = kwargs.pop("statement_timeout", None)
-        if statement_timeout is not None:
-            settings["max_execution_time"] = int(statement_timeout)
-        # Merge any user-supplied ``settings`` from kwargs into the local dict
-        # *before* applying the rest, otherwise ``client_kwargs.update(kwargs)``
-        # below would clobber the statement_timeout-derived max_execution_time.
-        extra_settings = kwargs.pop("settings", None)
-        if extra_settings:
-            settings.update(extra_settings)
-
-        client_kwargs = {
-            "host": info.host,
-            "port": int(info.port),
-            "database": info.database,
-            "username": info.user,
-            "password": info.password.get_secret_value() if info.password else "",
-            "secure": info.secure,
-            "settings": settings,
-        }
-        client_kwargs.update(kwargs)
-        return clickhouse_connect.get_client(**client_kwargs)
-
-    @classmethod
-    def get_mssql_connection(cls, info: MSSqlConnectionInfo):
-        return cls._connect_mssql_pyodbc(
-            host=info.host,
-            port=info.port,
-            database=info.database,
-            user=info.user,
-            password=info.password.get_secret_value() if info.password else None,
-            driver=info.driver,
-            kwargs={
-                "TDS_Version": info.tds_version,
-                **(info.kwargs if info.kwargs else {}),
-            },
-        )
-
-    @classmethod
-    def get_mssql_connection_from_url(
-        cls, connection_url: str, base_kwargs: dict[str, Any] | None = None
-    ):
-        parsed = urlparse(connection_url)
-        if parsed.scheme != "mssql":
-            raise WrenError(
-                ErrorCode.INVALID_CONNECTION_INFO,
-                "Invalid connection URL for MSSQL",
-            )
-
-        if not parsed.hostname or not parsed.path or not parsed.username:
-            raise WrenError(
-                ErrorCode.INVALID_CONNECTION_INFO,
-                "MSSQL connection URL must include user, host and database",
-            )
-
-        kwargs = dict(base_kwargs) if base_kwargs else {}
-        # parse_qsl already URL-decodes values, but we re-apply unquote_plus
-        # below only to components urlparse leaves encoded (user, path, password).
-        for key, value in urllib.parse.parse_qsl(parsed.query):
-            kwargs[key] = value
-        driver = kwargs.pop("driver", "ODBC Driver 18 for SQL Server")
-
-        return cls._connect_mssql_pyodbc(
-            host=parsed.hostname,
-            port=str(parsed.port or 1433),
-            database=unquote_plus(parsed.path.lstrip("/")),
-            user=unquote_plus(parsed.username),
-            password=unquote_plus(parsed.password) if parsed.password else None,
-            driver=driver,
-            kwargs=kwargs,
-        )
-
-    @staticmethod
-    def _connect_mssql_pyodbc(
-        host: str,
-        port: str,
-        database: str,
-        user: str | None,
-        password: str | None,
-        driver: str,
-        kwargs: dict[str, Any] | None = None,
-    ):
-        if pyodbc is None:  # pragma: no cover
-            raise WrenError(
-                ErrorCode.GET_CONNECTION_ERROR, "pyodbc is required for MSSQL"
-            )
-
-        connect_kwargs = dict(kwargs) if kwargs else {}
-        statement_timeout = connect_kwargs.pop("statement_timeout", None)
-        # Validate statement_timeout before opening the connection so a bad
-        # value can't leak the pyodbc connection we'd otherwise open first.
-        if statement_timeout is not None:
-            try:
-                statement_timeout = int(statement_timeout)
-            except (TypeError, ValueError) as exc:
-                raise WrenError(
-                    ErrorCode.INVALID_CONNECTION_INFO,
-                    f"Invalid statement_timeout for MSSQL: {statement_timeout!r}",
-                ) from exc
-
-        connection_parts = [
-            f"DRIVER={DataSourceExtension._escape_odbc_value(driver)}",
-            f"SERVER={host},{port}",
-            f"DATABASE={DataSourceExtension._escape_odbc_value(database)}",
-        ]
-        if user is None and password is None:
-            connection_parts.append("Trusted_Connection=yes")
-        elif user is None or password is None:
-            raise WrenError(
-                ErrorCode.INVALID_CONNECTION_INFO,
-                "MSSQL connection requires both user and password, "
-                "or neither (for Trusted_Connection)",
-            )
-        else:
-            connection_parts.append(
-                f"UID={DataSourceExtension._escape_odbc_value(user)}"
-            )
-            connection_parts.append(
-                f"PWD={DataSourceExtension._escape_odbc_value(password)}"
-            )
-
-        for key, value in connect_kwargs.items():
-            connection_parts.append(
-                f"{key}={DataSourceExtension._escape_odbc_value(str(value))}"
-            )
-
-        connection = pyodbc.connect(";".join(connection_parts))
-        DataSourceExtension._register_mssql_output_converters(connection)
-
-        if statement_timeout is not None:
-            connection.timeout = statement_timeout
-
-        return connection
-
-    @staticmethod
-    def _escape_odbc_value(value: str) -> str:
-        return "{" + value.replace("}", "}}") + "}"
-
-    @staticmethod
-    def _register_mssql_output_converters(connection) -> None:
-        connection.add_output_converter(
-            MSSQL_DATETIMEOFFSET_TYPE_CODE,
-            DataSourceExtension._decode_mssql_datetimeoffset,
-        )
-
-    @staticmethod
-    def _decode_mssql_datetimeoffset(value: bytes | None) -> dtlib.datetime | None:
-        if value is None:
-            return None
-        if len(value) != 20:
-            raise ValueError(
-                "unexpected mssql datetimeoffset payload length: "
-                f"expected 20, got {len(value)}"
-            )
-
-        year = int.from_bytes(value[0:2], "little")
-        month = int.from_bytes(value[2:4], "little")
-        day = int.from_bytes(value[4:6], "little")
-        hour = int.from_bytes(value[6:8], "little")
-        minute = int.from_bytes(value[8:10], "little")
-        second = int.from_bytes(value[10:12], "little")
-        nanoseconds = int.from_bytes(value[12:16], "little")
-        offset_hours = int.from_bytes(value[16:18], "little", signed=True)
-        offset_minutes = int.from_bytes(value[18:20], "little", signed=True)
-
-        tzinfo = dtlib.timezone(
-            dtlib.timedelta(hours=offset_hours, minutes=offset_minutes)
-        )
-        return dtlib.datetime(
-            year=year,
-            month=month,
-            day=day,
-            hour=hour,
-            minute=minute,
-            second=second,
-            microsecond=nanoseconds // 1000,
-            tzinfo=tzinfo,
-        )
-
-    @classmethod
-    def get_mysql_connection(cls, info: MySqlConnectionInfo) -> "MySQLdb.Connection":
-        import MySQLdb  # noqa: PLC0415
-
-        from wren.connector.mysql import _build_mysql_connect_kwargs  # noqa: PLC0415
-
-        return MySQLdb.connect(**_build_mysql_connect_kwargs(info))
-
-    @classmethod
-    def get_doris_connection(cls, info: DorisConnectionInfo) -> "MySQLdb.Connection":
-        import MySQLdb  # noqa: PLC0415
-
-        from wren.connector.mysql import _build_doris_connect_kwargs  # noqa: PLC0415
-
-        return MySQLdb.connect(**_build_doris_connect_kwargs(info))
-
-    @staticmethod
-    def get_postgres_connection(
-        info: PostgresConnectionInfo,
-    ) -> "psycopg.Connection":
-        """Open a native psycopg3 connection to PostgreSQL.
-
-        Returned object is a ``psycopg.Connection`` — the postgres connector
-        uses raw cursors and an OID-to-Arrow mapping to convert results.
-        """
-        import psycopg  # noqa: PLC0415
-
-        kwargs: dict[str, Any] = dict(info.kwargs) if info.kwargs else {}
-        return psycopg.connect(
-            host=info.host,
-            port=int(info.port),
-            dbname=info.database,
-            user=info.user,
-            password=(info.password and info.password.get_secret_value()),
-            **kwargs,
-        )
-
-    @staticmethod
-    def get_snowflake_connection(info: SnowflakeConnectionInfo):
-        from wren.connector.snowflake import make_snowflake_connection  # noqa: PLC0415
-
-        return make_snowflake_connection(info)
-
-    @staticmethod
-    def get_trino_connection(info: TrinoConnectionInfo):
-        """Return a ``trino.dbapi.Connection`` (not an ibis backend).
-
-        The wren SDK calls into ``wren.connector.trino`` for trino, which
-        operates directly on the DB-API cursor. ``get_connection`` only ever
-        re-routes here when something outside the v4 connector code path
-        explicitly requests a trino connection.
-        """
-        from trino.auth import BasicAuthentication  # noqa: PLC0415
-        from trino.dbapi import connect as trino_connect  # noqa: PLC0415
-
-        kwargs = dict(info.kwargs) if info.kwargs else {}
-        password = info.password.get_secret_value() if info.password else None
-
-        connect_kwargs: dict = {
-            "host": info.host,
-            "port": int(info.port),
-            "user": info.user,
-            "catalog": info.catalog,
-            "schema": info.trino_schema,
-            # Pin to UTC so timestamp-with-tz casts are deterministic.
-            "timezone": "UTC",
-        }
-        if info.user and password:
-            connect_kwargs["auth"] = BasicAuthentication(info.user, password)
-            connect_kwargs["http_scheme"] = "https"
-        connect_kwargs.update(kwargs)
-        return trino_connect(**connect_kwargs)
